feat: consistent colors in graph by sorting the data
The Time Series visualisation chooses colors for each series based on
the order they are returned. The order previously relied on the API
requests, which meant that it was random every run, causing random
colors to be assigned every time the graph was refreshed.

We now sort the Frames by the resource ID first and then by the series
name. This should give a consistent order as long as no new resources
are added. And even then, new resources usually have a higher ID and do
not break the current sorting, so only new labels can cause "issues".

Closes #7